backend/api/requirements.txt
flask
flask_cors
google.generativeai== 0.8.3
